# api-documentation.yaml
openapi: 3.0.0
info:
  title: PROgress Tracker API
  description: Breast Cancer Patient Survey Analytics Platform API
  version: 1.0.0
  contact:
    name: Support Team
    email: support@progresstracker.ca
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.progresstracker.ca
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        data:
          type: object
...
    Patient:
      type: object
      properties:
        patientId:
          type: string
        totalEvents:
          type: integer
        firstEvent:
          type: string
          format: date-time
        lastEvent:
          type: string
          format: date-time
        uniqueEvents:
          type: integer
        avgSatisfaction:
          type: number
        daysInactive:
          type: integer
...
paths:
...
  /api/patients:
    get:
      tags:
        - Patients
      summary: Get patients list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        200:
          description: Patients list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  patients:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    type: object
  /api/patients/{patientId}/journey:
    get:
      tags:
        - Patients
      summary: Get patient journey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Patient journey retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/patients/{patientId}/analytics:
    get:
      tags:
        - Patients
      summary: Get patient analytics
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Patient analytics retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/patients/{patientId}/notes:
    get:
      tags:
        - Patients
      summary: Get patient notes
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Patient notes retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

    post:
      tags:
        - Patients
      summary: Add patient note
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: patientId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - note
              properties:
                note:
                  type: string
      responses:
        200:
          description: Note added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  # Health Endpoints
...